Updated the CachePurge function to only pass directories to the deleteRecursive function when the file or directory actually exists.

+++ trunk/modules/Mortar/actions/CachePurge.class.php
@@ -19,7 +19,10 @@ class MortarActionCachePurge extends ActionBase
 		$clearPath[] = $config['path']['temp'] . 'loginTracker.sqlite';
 
 		foreach($clearPath as $path)
-			FileSystem::deleteRecursive($path);
+		{
+			if(is_dir($path))
+				FileSystem::deleteRecursive($path);
+		}
 
 		return true;
 	}
